fix(aw3): close jsonl identity and client line cap
