+++ crates/atm/tests/fixtures/log-jsonl-default.json
@@ -1,0 +1,14 @@
+{
+  "records": [
+    {
+      "timestamp": "1970-01-01T00:00:00.100Z",
+      "level": "info",
+      "service": "atm",
+      "target": "cli.command",
+      "action": "send",
+      "message": "jsonl fixture",
+      "fields": {}
+    }
+  ],
+  "truncated": false
+}
